requirements.txt
alembic==1.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
nltk==3.4.5
numpy==1.17.1
oauthlib==3.1.0
pandas==0.24.2
pkg-resources==0.0.0
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.8
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.6
WTForms==2.2.1
